修改locker的清空机制

+++ YTools/system/locker/lock.py
@@ -9,9 +9,10 @@ class LockerClient:
 
 	NAME = "YLocker"
 	VALFILE = "val.pkl"
+	EMPTY_FLAG = "_EMPTY"
 
 	def __init__(self):
-		self.fakepath = new_fakefolder(self.NAME)
+		pass
 
 	def getfolder(self , foldername):
 		# 给定目录名，返回其绝对路径
@@ -46,10 +47,10 @@ class LockerClient:
 		
 	def remove(self , key):
 		path = self.ensurefile(key)
-		with FileLock(self.getfolder(key)): #锁定然后删除文件
+		with FileLock(self.getfolder(key)): #锁定然后写入
 			os.remove(path)
 		return True
-	
+
 	def plus(self , key , val = 1):
 		'''同步加一，返回加后的值'''		
 		path = self.ensurefile(key)
@@ -74,15 +75,61 @@ class LockerClient:
 
 		return True
 
-	def ask_prefix(self , prefix):
-		'''查询所有key的前缀是给定prfix的key'''
-		
-		all_dirs = os.listdir(self.fakepath)
-		return [ p for p in all_dirs if P.isdir(p) and p.startswith(prefix) ]
+	def _list_folders(self , folder):
+		'''列出所目标子文件夹'''
+		ret = []
+		subfolers = os.listdir(folder)
+		if self.VALFILE in subfolers: #存在值文件的是一个目标节点
+			ret.append(folder)
+		for p in subfolers:
+			p = P.join(folder,p)
+			if P.isdir(p):
+				ret = ret + self._list_folders(p)
+		return ret
+
+	def ask_prefix(self , prefix , return_abs = False , only_suffix = False , not_none = False):
+		'''查询所有key的前缀是给定prfix的key
+		not_none：只返回非None的元素
+		'''
+		path = self.getfolder(prefix)
+		ret = self._list_folders(path) #询问前缀下面的所有文件夹
+		if not_none:
+			ret = [x for x in ret if self.get(x) is not None]
+		if return_abs:
+			return ret
+		if only_suffix:
+			ret = [P.relpath(x , path) for x in ret] #求相对与前缀的相对路径
+		else:
+			the_path = self.getfolder("")
+			ret = [P.relpath(x , the_path) for x in ret] #求伪路径
+		return ret
 
 	def clear(self):
 		'''清除所有key'''
-		all_dirs = self.ask_prefix("")
+		all_dirs = self.ask_prefix("" , return_abs = True)
 		for p in all_dirs:
-			os.removedirs( p )
-		return True 
+			self.remove(p)
+
+			if len( os.listdir(p) ) == 0: #为空才删除文件夹，否则就保留
+				os.removedirs( p )
+		self.clear_dir()
+		return True
+
+	def clear_dir(self , path = None):
+		'''清除所有空文件夹'''
+		if path is None:
+			path = self.getfolder("")
+		flag = True #自己是否能清除
+		all_dirs = [P.join(path,p) for p in os.listdir(path) ]
+		for p in all_dirs:
+			if P.isdir(p):
+				flag = flag and self.clear_dir(p)
+			else:
+				flag = False
+		if flag: #现在应该已经没有文件了
+			try:
+				os.removedirs( path )
+			except Exception:
+				flag = False #不让删就算了
+		return flag
+
